Extract randomized test infrastructure for use in other tests
